+++ src/mdf_viewer/view/widgets/__init__.py
@@ -1,6 +1,14 @@
 """Small reusable view widgets."""
 
 from mdf_viewer.view.widgets.color_swatch import ColorSwatch
+from mdf_viewer.view.widgets.icons import _icon_suffix, _load_icon
 from mdf_viewer.view.widgets.splitter import make_splitter
+from mdf_viewer.view.widgets.visibility_toggle_button import VisibilityToggleButton
 
-__all__ = ["ColorSwatch", "make_splitter"]
+__all__ = [
+    "ColorSwatch",
+    "VisibilityToggleButton",
+    "_icon_suffix",
+    "_load_icon",
+    "make_splitter",
+]